requirements.txt
# 智能客服系统 - Python 依赖包
# 核心框架
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3

# 数据库
SQLAlchemy==2.0.23
alembic==1.13.1

# 安全
Werkzeug==3.0.1

# 任务调度
APScheduler==3.10.4

# 日志
loguru==0.7.2

# 图像处理
Pillow==10.1.0
pyautogui==0.9.54

# Windows 系统操作
pywin32==306

# HTTP 请求
requests==2.31.0

# 向量搜索和AI
numpy>=1.26.0
scikit-learn>=1.3.2

# 可选：OCR 功能（如果需要）
# paddlepaddle==2.5.2
# paddleocr==2.7.3

# 开发工具
pytest==7.4.3
pytest-flask==1.3.0

# 系统监控
psutil==5.9.6

# Excel处理
pandas==2.1.4
openpyxl==3.1.2
